pyproject.toml
[project]
name = "athar-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "llama-index>=0.14.20",
    "llama-index-embeddings-openai>=0.6.0",
    "llama-index-llms-openai>=0.7.5",
    "pydantic>=2.13.2",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
    "uvicorn>=0.44.0",
    "wikipedia-api>=0.14.1",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
